Renamed LevelGenerator singleton to just Level, and former Level to TileBatch.
