+++ src/game-loop/src/components/PhysicsComponent.cpp
@@ -1,6 +1,6 @@
 #include "components/PhysicsComponent.hpp"
 #include "main-dude/MainDude.hpp"
-#include "LevelGenerator.hpp"
+#include "Level.hpp"
 #include "Collisions.hpp"
 
 // Using C-style <math.h> instead of <cmath> because of some symbols (namely std::copysign)
@@ -75,7 +75,7 @@ void PhysicsComponent::update(MainDude &main_dude, uint32_t delta_time_ms)
         if (!initial_check_done)
         {
             MapTile *neighbours[9] = {nullptr};
-            collisions::get_neighbouring_tiles(LevelGenerator::instance().getLevel(), _position.x, _position.y, neighbours);
+            collisions::get_neighbouring_tiles(Level::instance().get_tile_batch(), _position.x, _position.y, neighbours);
             const auto *overlapping_tile = collisions::overlaps_strict(neighbours, _position.x, _position.y, _dimensions.width, _dimensions.height);
             if (overlapping_tile)
             {
@@ -129,7 +129,7 @@ void PhysicsComponent::update(MainDude &main_dude, uint32_t delta_time_ms)
                 // Step on X axis
 
                 _position.x += copysign(smallest_position_step, temp_velocity_x);
-                collisions::get_neighbouring_tiles(LevelGenerator::instance().getLevel(), _position.x, _position.y, neighbours);
+                collisions::get_neighbouring_tiles(Level::instance().get_tile_batch(), _position.x, _position.y, neighbours);
                 const auto* overlapping_tile = collisions::overlaps(neighbours, _position.x, _position.y, _dimensions.width, _dimensions.height);
                 if (overlapping_tile)
                 {
@@ -161,7 +161,7 @@ void PhysicsComponent::update(MainDude &main_dude, uint32_t delta_time_ms)
                 // Step on Y axis
 
                 _position.y += copysign(smallest_position_step, temp_velocity_y);
-                collisions::get_neighbouring_tiles(LevelGenerator::instance().getLevel(), _position.x,  _position.y, neighbours);
+                collisions::get_neighbouring_tiles(Level::instance().get_tile_batch(), _position.x,  _position.y, neighbours);
                 const auto* overlapping_tile = collisions::overlaps(neighbours, _position.x, _position.y, _dimensions.width, _dimensions.height);
                 if (overlapping_tile)
                 {
