+++ src/level-generator/generator/CMakeLists.txt
@@ -3,16 +3,16 @@ project(LevelGenerator)
 # FIXME: This project is very poorly written (copy-pasted from Spelunky-NDS), needs refactoring.
 
 add_library(LevelGenerator STATIC
-        src/LevelGenerator.cpp
-        interface/LevelGenerator.hpp
+        src/Level.cpp
+        interface/Level.hpp
         interface/MapTile.hpp
         interface/MapTileType.hpp
         src/MapTile.cpp
         interface/SplashScreenType.hpp
         interface/TileOrientation.hpp
         interface/Direction.hpp
-        interface/Level.hpp
-        src/Level.cpp
+        interface/TileBatch.hpp
+        src/TileBatch.cpp
 )
 
 target_include_directories(LevelGenerator
